nova_agents/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
nova-act
strands-agents>=1.2.0
strands-agents-tools
strands-nova